% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR.TSHT.R
\name{analysis.CI}
\alias{analysis.CI}
\title{Handle the possible union of CIs}
\usage{
analysis.CI(CI.matrix, true.val, grid.size)
}
\arguments{
\item{CI.matrix}{a matrix, where each row represents a CI}

\item{true.val}{a numeric scalar indecating the true value of treatment effect}

\item{grid.size}{a numeric scalar indicating the size of the grid}
}
\value{
\item{\code{CI.coverage}}{a numeric scalar denoting whether the CI contains the true value}
    \item{\code{CI.length}}{a numeric scalar denoting the length of the CI.}
    \item{\code{grid.seq}}{a numeric vector indicating the sequence of grids in CI.matrix}
}
\description{
Take union of CIs accoding to the specific grid sizes
}
