% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR.TSHT.R
\name{normalize}
\alias{normalize}
\title{Normalize function}
\usage{
normalize(x)
}
\arguments{
\item{x}{a numeric vector of input data}
}
\value{
a numeric vector of standardized output data
}
\description{
Standardization of a vector to be 0 mean and unit variance
}
\examples{
x <- rnorm(10, 2, 5);
y <- normalize(x);
}
