% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR.TSHT.R
\name{Searching.Sampling.sumstats}
\alias{Searching.Sampling.sumstats}
\title{Searhing-and-Sampling method}
\usage{
Searching.Sampling.sumstats(
  ITT_Y,
  ITT_D,
  V_Gamma,
  V_gamma,
  VHat,
  n1,
  n2,
  beta.grid = NULL,
  rho = NULL,
  S = 1000
)
}
\arguments{
\item{ITT_Y}{a numeric vector of GWAS summary statistics of the outcome}

\item{ITT_D}{a numeric vector of GWAS summary statistics of the treatment}

\item{VHat}{a numeric vector denoting the set of valid and relevant IVs, usually obtained in the results of TSHT.sum}

\item{n1}{the sample size of GWAS summary statistics  of the outcome}

\item{n2}{the sample size of GWAS summary statistics  of the treatment}

\item{beta.grid}{a numeric vector of the grid for sampling}

\item{rho}{a numeric scalar denoting thresholding level used in the sampling property}

\item{S}{a positive integer indicating the number of bootstrap resampling for computing the confidence interval, with default 1000}

\item{SE_Y}{a numeric vector of standard errors of ITT_Y}

\item{SE_D}{a numeric vector of standard errors of ITT_D}
}
\value{
\item{\code{CI.union}}{a two dimensional numeric vector denoting the 1-alpha confidence intervals for betaHat with lower and upper endpoints constructed by Searching-Sampling method.}
    \item{\code{rule}}{a boolean scalar denoting whether the identification condition is satisfied or not.}
    \item{\code{CI}}{a numeric matrix denoting the confidence intervals for betaHat constructed by valid candidates for betaHat.}
}
\description{
Get a possible confidence interval using searching method
}
