% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR.TSHT.R
\name{Searching.sumstats}
\alias{Searching.sumstats}
\title{Searhing method}
\usage{
Searching.sumstats(
  ITT_Y,
  ITT_D,
  V_Gamma,
  V_gamma,
  VHat,
  n1,
  n2,
  beta.grid = NULL
)
}
\arguments{
\item{ITT_Y}{a numeric vector of GWAS summary statistics of the outcome}

\item{ITT_D}{a numeric vector of GWAS summary statistics of the treatment}

\item{VHat}{a numeric vector denoting the set of valid and relevant IVs, usually obtained in the results of TSHT.sum}

\item{n1}{the sample size of GWAS summary statistics  of the outcome}

\item{n2}{the sample size of GWAS summary statistics  of the treatment}

\item{beta.grid}{a numeric vector of the grid for sampling}

\item{SE_Y}{a numeric vector of standard errors of ITT_Y}

\item{SE_D}{a numeric vector of standard errors of ITT_D}
}
\value{
\item{\code{CI.search}}{a two dimensional numeric vector denoting the 1-alpha confidence intervals for betaHat with lower and upper endpoints constructed by Searching method.}
    \item{\code{rule}}{a boolean scalar denoting whether the identification condition is satisfied or not}
    \item{\code{valid.grid}}{a numeric vector denoting the candidates satisfying certain thresholding condition}
}
\description{
Get a possible confidence interval using searching method
}
