% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR.TSHT.R
\name{SearchingSampling.sumstats}
\alias{SearchingSampling.sumstats}
\title{Searhing-and-Sampling method}
\usage{
SearchingSampling.sumstats(
  ITT_D,
  ITT_Y,
  SE_D,
  SE_Y,
  n1,
  n2,
  CI.init = NULL,
  a = 0.6,
  Sampling = TRUE,
  rho = NULL,
  M = 1000,
  prop = 0.1
)
}
\arguments{
\item{ITT_D}{a numeric vector of GWAS summary statistics of the treatment}

\item{ITT_Y}{a numeric vector of GWAS summary statistics of the outcome}

\item{SE_D}{a numeric vector of standard errors of ITT_D}

\item{SE_Y}{a numeric vector of standard errors of ITT_Y}

\item{n1}{the sample size of GWAS summary statistics  of the treatment}

\item{n2}{the sample size of GWAS summary statistics  of the outcome}

\item{CI.init}{initial interval for beta. If \code{NULL}, it will be generated automatically. (default=\code{NULL})}

\item{a}{grid size for constructing beta grids (default=0.6)}

\item{Sampling}{if \code{TRUE}, use the proposed sampling method; else use the proposed searching method. (default=\code{TRUE})}

\item{rho}{a numeric scalar denoting thresholding level used in the sampling property}

\item{M}{sampling times. (default=1000)}

\item{prop}{proportion of intervals kept when sampling. (default=0.1)}
}
\value{
\item{\code{CI}}{a numeric matrix denoting the confidence intervals for betaHat constructed by valid candidates for betaHat.}
    \item{\code{rule}}{a boolean scalar denoting whether the identification condition is satisfied or not.}
    \item{\code{VHat}}{valid instruments}
    \item{\code{CI.init}}{initial confidence interval for searching and sampling}
    \item{\code{TSHT.out}}{standard MR-TSHT output}
}
\description{
Get a possible confidence interval using searching method
}
