% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR.TSHT.R
\name{TSHT.sumstats}
\alias{TSHT.sumstats}
\title{TSHT using summary statistics}
\usage{
TSHT.sumstats(
  ITT_Y,
  ITT_D,
  SE_Y,
  SE_D,
  n1,
  n2,
  tuning = 1,
  max_clique = FALSE,
  alpha = 0.05
)
}
\arguments{
\item{ITT_Y}{a numeric vector of GWAS summary statistics of the outcome}

\item{ITT_D}{a numeric vector of GWAS summary statistics of the treatment}

\item{SE_Y}{a numeric vector of standard errors of ITT_Y}

\item{SE_D}{a numeric vector of standard errors of ITT_D}

\item{n1}{the sample size of GWAS summary statistics  of the outcome}

\item{n2}{the sample size of GWAS summary statistics  of the treatment}

\item{tuning}{a numeric scalar value tuning parameter for TSHT, with default 1}

\item{max_clique}{an option to replace the majority and plurality voting procedures with finding maximal clique in the IV voting matrix, with default FALSE}

\item{alpha}{a numeric scalar value between 0 and 1 indicating the significance level for the confidence interval, with default 0.05}
}
\value{
\item{\code{VHat}}{a numeric vector denoting the set of valid and relevant IVs}
    \item{\code{betaHat}}{a numeric scalar denoting the estimate of treatment effect.}
    \item{\code{seHat}}{a numeric scalar denoting the estimated standard error of betaHat.}
    \item{\code{ci}}{a two dimensional numeric vector denoting the 1-alpha confidence intervals for betaHat with lower and upper endpoints.}
}
\description{
Conduct two-stage hard thresholding using summary statistics
}
