% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR.TSHT.R
\name{var_TSHT}
\alias{var_TSHT}
\title{Variance of TSHT estimator}
\usage{
var_TSHT(gamma, Gamma, V_gamma, V_Gamma)
}
\arguments{
\item{gamma}{a numeric vector of marginal coefficients from IVs to treatment}

\item{Gamma}{a numeric vector of marginal coefficients from IVs to outcome}

\item{V_gamma}{a numeric covariance matrix of gamma}

\item{V_Gamma}{a numeric covariance matrix of Gamma}
}
\value{
a numeric scalar value indicating the variance of the estimator
}
\description{
Calculate the variance of TSHT estimator after getting V_Hat
}
